Refactored lint/settings
